# Conflicts:
#	src/main/java/org/nurim/nurim/Controller/AuthController.java
#	src/main/java/org/nurim/nurim/Controller/MemberController.java
#	src/main/java/org/nurim/nurim/config/WebConfig.java
#	src/main/java/org/nurim/nurim/config/auth/LoginFilter.java
#	src/main/java/org/nurim/nurim/config/auth/LoginSuccessHandler.java
#	src/main/java/org/nurim/nurim/domain/dto/LoginResponse.java
#	src/main/java/org/nurim/nurim/domain/dto/policy/PlolicyDto.java
#	src/main/java/org/nurim/nurim/domain/entity/Member.java
#	src/main/java/org/nurim/nurim/domain/entity/api/ChildCare.java
#	src/main/java/org/nurim/nurim/service/CallApiService.java
#	src/main/java/org/nurim/nurim/service/PrincipalDetailsService.java

+++ src/main/java/org/nurim/nurim/service/MemberService.java
@@ -3,33 +3,42 @@ package org.nurim.nurim.service;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
+import org.nurim.nurim.AmazonS3.FileUploadService;
 import org.nurim.nurim.domain.dto.member.*;
 import org.nurim.nurim.domain.entity.Member;
 import org.nurim.nurim.domain.entity.MemberImage;
+import org.nurim.nurim.domain.entity.MemberRole;
 import org.nurim.nurim.repository.MemberImageRepository;
 import org.nurim.nurim.repository.MemberRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Optional;
 
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 @Log4j2
 public class MemberService {
-    private static final String DEFAULT_PROFILE_IMAGE_URL = "https://i.stack.imgur.com/l60Hf.png";
+//    private static final String DEFAULT_PROFILE_IMAGE_URL = "https://i.stack.imgur.com/l60Hf.png";
 
     private final MemberRepository memberRepository;
     private final MemberImageRepository memberImageRepository;
+    private final FileUploadService fileUploadService;
 
     @Autowired
     private final PasswordEncoder passwordEncoder;
 
-    // 회원 정보 입력
+    // 일반 회원 가입
     @Transactional
     public CreateMemberResponse createMember(CreateMemberRequest request) {
 
@@ -40,16 +49,21 @@ public class MemberService {
             throw new DataIntegrityViolationException("이미 존재하는 회원입니다.");   // 전역예외처리 필요
         }
 
+        // 초기 프로필 이미지 URL 설정 (S3 버킷에 저장된 기본 이미지 URL)
+        String defaultProfileImageUrl = "https://nurimplus1.s3.ap-northeast-2.amazonaws.com/default-image.jpg";
+
+        // 초기 필드값 default 설정
         Member member = Member.builder()
                 .memberEmail(request.getMemberEmail())
                 .memberPw(passwordEncoder.encode(request.getMemberPw()))
                 .memberNickname(request.getMemberNickname())
-                .memberAge(request.getMemberAge())
-                .gender(request.isGender())
-                .memberResidence(request.getMemberResidence())
-                .memberMarriage(request.isMemberMarriage())
-                .memberIncome(request.getMemberIncome())
-                .type(request.isType())
+                .memberAge(30)
+                .gender(true)
+                .memberResidence("거주지 주소를 입력해주세요.")
+                .memberMarriage(true)
+                .memberIncome("소득 정보를 입력해주세요.")
+                .type(false)
+                .memberRole(MemberRole.USER)
                 .build();
 
         Member savedMember = memberRepository.save(member);
@@ -57,7 +71,7 @@ public class MemberService {
         // 기본 이미지 경로 MemberImage에 설정하여 저장
         MemberImage memberImage = new MemberImage();
         memberImage.setMember(savedMember);
-        memberImage.setMemberProfileImage(DEFAULT_PROFILE_IMAGE_URL); // 정적 경로 참조
+        memberImage.setMemberProfileImage(defaultProfileImageUrl); // 정적 경로 참조
         memberImageRepository.save(memberImage);
 
         // 회원 정보에 이미지 정보 연결
@@ -73,9 +87,113 @@ public class MemberService {
                 savedMember.getMemberResidence(),
                 savedMember.isMemberMarriage(),
                 savedMember.getMemberIncome(),
-                savedMember.isType());
+                savedMember.isType(),
+                savedMember.getMemberRole(),
+                savedMember.getMemberProfileImage()
+        );
 
     }
+
+    // 관리자 회원 가입
+    @Transactional
+    public CreateMemberResponse createAdmin(CreateMemberRequest request) {
+
+        // 회원 정보 유효성 검증 (클래스 별도 생성 예정)
+        // validateMemberRequest(request);
+
+        if (memberRepository.findMemberByMemberEmail(request.getMemberEmail()).isPresent()) {
+            throw new DataIntegrityViolationException("이미 존재하는 회원입니다.");   // 전역예외처리 필요
+        }
+
+        // 초기 프로필 이미지 URL 설정 (S3 버킷에 저장된 기본 이미지 URL)
+        String defaultProfileImageUrl = "https://nurimplus1.s3.ap-northeast-2.amazonaws.com/default-image.jpg";
+
+        // 초기 필드값 default 설정
+        Member member = Member.builder()
+                .memberEmail(request.getMemberEmail())
+                .memberPw(passwordEncoder.encode(request.getMemberPw()))
+                .memberNickname(request.getMemberNickname())
+                .memberAge(30)
+                .gender(true)
+                .memberResidence("서울시 강남구 강남대로 405")
+                .memberMarriage(true)
+                .memberIncome("해당 없음")
+                .type(false)
+                .memberRole(MemberRole.ADMIN)
+                .build();
+
+        Member savedMember = memberRepository.save(member);
+
+        // 기본 이미지 경로 MemberImage에 설정하여 저장
+        MemberImage memberImage = new MemberImage();
+        memberImage.setMember(savedMember);
+        memberImage.setMemberProfileImage(defaultProfileImageUrl); // 정적 경로 참조
+        memberImageRepository.save(memberImage);
+
+        // 회원 정보에 이미지 정보 연결
+        savedMember.setMemberImage(memberImage);
+        memberRepository.save(savedMember);
+
+        return new CreateMemberResponse(savedMember.getMemberId(),
+                savedMember.getMemberEmail(),
+                savedMember.getMemberPw(),
+                savedMember.getMemberNickname(),
+                savedMember.getMemberAge(),
+                savedMember.isGender(),
+                savedMember.getMemberResidence(),
+                savedMember.isMemberMarriage(),
+                savedMember.getMemberIncome(),
+                savedMember.isType(),
+                savedMember.getMemberRole(),
+                savedMember.getMemberProfileImage()
+        );
+
+    }
+
+    // 회원 정보 입력
+    @Transactional
+    public CreateMemberResponse createMemberInfo(CreateMemberInfoRequest request) {
+
+        Member member = Member.builder()
+                .memberEmail(getMember().getMemberEmail())
+                .memberPw(getMember().getMemberPw())
+                .memberNickname(getMember().getMemberNickname())
+                .memberAge(request.getMemberAge())
+                .gender(request.isGender())
+                .memberResidence(request.getMemberResidence())
+                .memberMarriage(request.isMemberMarriage())
+                .memberIncome(request.getMemberIncome())
+                .type(request.isType())
+                .memberRole(getMember().getMemberRole())
+                .build();
+
+//        Member savedMember = memberRepository.save(member);
+//
+//        // 기본 이미지 경로 MemberImage에 설정하여 저장
+//        MemberImage memberImage = new MemberImage();
+//        memberImage.setMember(savedMember);
+//        memberImage.setMemberProfileImage(DEFAULT_PROFILE_IMAGE_URL); // 정적 경로 참조
+//        memberImageRepository.save(memberImage);
+//
+//        // 회원 정보에 이미지 정보 연결
+//        savedMember.setMemberImage(memberImage);
+//        memberRepository.save(savedMember);
+
+        return new CreateMemberResponse(member.getMemberId(),
+                member.getMemberEmail(),
+                member.getMemberPw(),
+                member.getMemberNickname(),
+                member.getMemberAge(),
+                member.isGender(),
+                member.getMemberResidence(),
+                member.isMemberMarriage(),
+                member.getMemberIncome(),
+                member.isType(),
+                member.getMemberRole(),
+                member.getMemberProfileImage());
+
+    }
+
 
     // 특정 회원 조회
     public ReadMemberResponse readMemberById(Long memberId) {
@@ -89,7 +207,7 @@ public class MemberService {
             profileimageUrl = foundMember.getMemberImage().getMemberProfileImage();
         } else {
             // 프로필 이미지가 등록되지 않은 경우
-            profileimageUrl = "기본 프로필 이미지 URL";
+            profileimageUrl = "https://nurimplus1.s3.ap-northeast-2.amazonaws.com/default-image.jpg";
         }
 
         String expertFileUrl;
@@ -125,12 +243,14 @@ public class MemberService {
         Member foundMember = memberRepository.findById(memberId)
                 .orElseThrow(() -> new EntityNotFoundException("😥해당 memberId로 조회된 회원 정보가 없습니다."));
 
-        String rawPw = request.getMemberPw();
-        String encPw = passwordEncoder.encode(rawPw);
+        // 현재 로그인한 사용자만 수정 가능
+        if(!foundMember.getMemberEmail().equals(getMember().getMemberEmail())) {
+            throw new AccessDeniedException("수정 권한이 없습니다.");
+        }
 
         // Member 정보 업데이트
         foundMember.update(
-                encPw,
+                passwordEncoder.encode(request.getMemberPw()),
                 request.getMemberNickname(),
                 request.getMemberAge(),
                 request.isGender(),
@@ -138,7 +258,7 @@ public class MemberService {
                 request.isMemberMarriage(),
                 request.getMemberIncome(),
                 request.isType());
-//
+
 //        // MemberImage 정보 업데이트
 //        String newMemberProfileImage = request.getMemberProfileImage(); // 새로운 이미지 정보
 //        UpdateMemberImageRequest imageRequest = new UpdateMemberImageRequest(newMemberProfileImage); // 이미지 정보 갖는 객체
@@ -169,36 +289,72 @@ public class MemberService {
         Member foundMember = memberRepository.findById(memberId)
                 .orElseThrow(() -> new EntityNotFoundException("😥해당 memberId로 조회된 회원 정보가 없습니다."));
 
+        // 현재 로그인한 사용자만 탈퇴 가능
+        if(!foundMember.getMemberEmail().equals(getMember().getMemberEmail())) {
+            throw new AccessDeniedException("이 계정 탈퇴에 대한 권한이 없습니다.");
+        }
+
         memberRepository.delete(foundMember);
 
         return new DeleteMemberResponse(foundMember.getMemberId());
 
     }
 
+
     // context에서 회원정보 가져오기
     public Member getMember() {
+
+        // SecurityContext에서 인증 정보 추출
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        if(authentication == null || !authentication.isAuthenticated()) {
+           log.info("인증 객체를 찾을 수 없습니다.");
+        }
 
         String username = authentication.getName();   // 사용자 이메일 추출
 
+        if(username == null) {
+            log.info("사용자 이메일 정보가 없습니다.");
+        }
+
+        log.info("😀사용자 이메일" + username);
+
         Member member = memberRepository.findMemberByMemberEmail(username)
-                .orElseThrow(() -> new EntityNotFoundException("사용자 정보를 찾을 수 없습니다."));
+                .orElseThrow(() -> new EntityNotFoundException("사용자 정보를 DB에서 찾을 수 없습니다."));
 
         return member;
     }
 
-    public Member findMemberByMemberEmail(String username) {
+
+    public Member readMemberByMemberEmail(String username) {
         Member foundMember = memberRepository.findMemberByMemberEmail(username)
-                .orElseThrow(() -> new EntityNotFoundException("해당 이메일로 회원을 찾을 수 없습니다."));
+                .orElseThrow(() -> new EntityNotFoundException("😥해당 이메일로 회원을 찾을 수 없습니다."));
 
         return foundMember;
     }
 
     public Member getMemberById(Long memberId) {
+
         Member foundMember = memberRepository.findById(memberId)
                 .orElseThrow(() -> new EntityNotFoundException("해당 memberId로 회원을 찾을 수 없습니다."));
 
+        log.info("😀"+foundMember);
+
         return foundMember;
+    }
+
+    public boolean isCurrentUser(Long memberId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication == null || !authentication.isAuthenticated()) {
+            return false; // 로그인한 사용자가 없는 경우
+        }
+
+        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+        String currentUsername = userDetails.getUsername();
+
+        // 현재 로그인한 사용자의 username과 memberId에 해당하는 회원의 username이 일치하는지 확인
+        Optional<Member> memberOptional = memberRepository.findById(memberId);
+        return memberOptional.isPresent() && memberOptional.get().getMemberEmail().equals(currentUsername);
     }
 
 }
